Fixed the Unheld Directional Issue
The probable was caused by the joystick being off center. The logic was built around ideal scenarios where the only off-perfect scenarios were caused by human error, but in practice the sticks turned out to be off of their true axis which caused the front wheels to believe the logic wanted it to go to a direction unintended by the user.